refactor : valid class에 대한 분리 - viewmodel안에 있던 valid를 분리처리
