+++ feature/login/src/main/kotlin/jinproject/stepwalk/login/screen/findpassword/FindPasswordScreen.kt
@@ -29,8 +29,8 @@ import jinproject.stepwalk.login.component.EmailVerificationField
 import jinproject.stepwalk.login.component.EnableButton
 import jinproject.stepwalk.login.component.IdDetail
 import jinproject.stepwalk.login.component.PasswordDetail
-import jinproject.stepwalk.login.screen.signup.SignValid
-import jinproject.stepwalk.login.screen.signup.ValidValue
+import jinproject.stepwalk.login.screen.state.SignValid
+import jinproject.stepwalk.login.screen.state.AccountValid
 import jinproject.stepwalk.login.screen.state.Verification
 import jinproject.stepwalk.login.utils.MAX_ID_LENGTH
 import jinproject.stepwalk.login.utils.MAX_PASS_LENGTH
@@ -60,7 +60,7 @@ internal fun FindPasswordScreen(
 @Composable
 private fun FindPasswordScreen(
     findAccountPassword: () -> FindAccountPassword,
-    valids : ValidValue,
+    valids : AccountValid,
     updateFindEvent : (FindEvent,String) -> Unit,
     requestEmailVerification : () -> Unit,
     requestFindAccount: () -> Unit,
@@ -161,7 +161,7 @@ private fun PreviewFindPasswordScreen(
 ) = StepWalkTheme {
     FindPasswordScreen(
         findAccountPassword = { FindAccountPassword(nextStep = true) },
-        valids = ValidValue(),
+        valids = AccountValid(),
         updateFindEvent = {_,_ ->},
         requestEmailVerification = {  },
         requestFindAccount = {  },
